refactor project folder structure
