+++ game.py
@@ -2,13 +2,13 @@
 import sys
 from os import path
 import pygame
-sys.path.insert(1, 'src/config')
-sys.path.insert(1, 'src/screens')
-sys.path.insert(1, 'src/assets')
+sys.path.insert(1, 'config')
+sys.path.insert(1, 'screens')
+sys.path.insert(1, 'assets')
 from config import HEIGHT, WIDTH, FPS, game, world_speed, TELA_INICIAL, QUIT, GAME, TELA_FINAL
-from src.screens.TelaInicio import tela_inicial
-from src.screens.TelaJogo import tela_jogo
-from src.screens.TelaFinal import tela_final
+from screens.TelaInicio import tela_inicial
+from screens.TelaJogo import tela_jogo
+from screens.TelaFinal import tela_final
 
 
 pygame.init()
