+++ screens/TelaInicio.py
@@ -2,8 +2,8 @@ import pygame
 import random
 from os import path
 import sys 
-sys.path.insert(1,'src/config')
-sys.path.insert(1, 'src/assets')
+sys.path.insert(1,'config')
+sys.path.insert(1, 'assets')
 
 from config import HEIGHT, WIDTH, FPS, game, world_speed, TELA_INICIAL, QUIT, GAME
 from assets import load_assets
